backend/projects/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
supabase==2.19.0
